backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
httpx==0.27.2
python-multipart==0.0.17
python-dotenv==1.0.1
